+++ R/bstfun-package.R
@@ -16,5 +16,6 @@ utils::globalVariables(c(".", "exponentiate"))
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
 #' @importFrom lifecycle deprecate_soft
+#' @importFrom lifecycle deprecated
 ## usethis namespace: end
 NULL